Send array of result positions to backend
Chang from sending a hard coded gold, silve and bronze position to
sending an array of results that contain place, score and an array of
players on it.  This makes it possible to send the players of teams and
shared placements for players that score the same in a match.